Consolidate tests of particle paraphernalia pickling
